Implement iOS media session and remote control

+++ androidApp/src/main/java/org/blackcandy/android/bridge/AlbumComponent.kt
@@ -37,7 +37,7 @@ class AlbumComponent(
 
     @Serializable
     data class AlbumData(
-        val albumId: Int,
-        val songId: Int?,
+        val albumId: Long,
+        val songId: Long?,
     )
 }
