+++ androidApp/src/main/java/org/blackcandy/android/bridge/SongsComponent.kt
@@ -43,6 +43,6 @@ class SongsComponent(
 
     @Serializable
     data class SongsData(
-        val songId: Int,
+        val songId: Long,
     )
 }
